Fix subnet initialization
