+++ tests/test_components.py
@@ -481,36 +481,6 @@ def test_inport_closed_propagation(net, discard):
     assert dis.values == ['000005']
 
 
-@requires_patch
-def test_subnet_with_substreams(net, discard):
-    # tracing = True
-    gen = net.add_component("Generate", GenSS, COUNT=15)
-    passnet = net.add_component("Subnet", PassthruNet)
-    dis = net.add_component("Discard", discard)
-
-    net.connect("Generate.OUT", "Subnet.IN")
-    net.connect("Subnet.OUT", "Discard.IN")
-
-    # net.reset()
-    # net._build_runners()
-    # net._open_ports()
-    # assert isinstance(passnet.ports['OUT'].component, Passthru)
-    # assert isinstance(passnet.ports['OUT'].sender, ComponentRunner)
-    # assert isinstance(passnet.ports['IN'].component, Passthru)
-    # assert isinstance(passnet.ports['IN'].receiver, ComponentRunner)
-
-
-    # FIXME: need a separate test for the NULL port behavior
-    # net.connect("Subnet.*SUBEND", "WTC.IN")
-    net.go()
-
-    assert dis.values == [
-        '', '000015', '000014', '000013', '000012', '000011', '',
-        '', '000010', '000009', '000008', '000007', '000006', '',
-        '', '000005', '000004', '000003', '000002', '000001', ''
-    ]
-
-
 def test_synced_receive1(net, discard):
     """
     components using `fn.synced` (Group) close all synced inports on the first
@@ -832,36 +802,6 @@ def test_network_export():
     assert len(net.outports.keys()) == 1
 
 
-def test_subnet_decorator():
-
-    @outport("OUT")
-    @inport("IN")
-    @subnet
-    def DecoratedPassNet(sub):
-        sub.add_component('Head', SlowPass, DELAY=0.01)
-        sub.add_component('Tail', SlowPass, DELAY=0.01)
-
-        sub.connect('Head.OUT', 'Tail.IN')
-
-        sub.export('Head.IN', 'IN')
-        sub.export('Tail.OUT', 'OUT')
-
-    net = Network()
-
-    gen = net.add_component("Generate", GenSS, COUNT=5)
-    passnet = net.add_component("Subnet", DecoratedPassNet)
-    dis = net.add_component("Discard", Discard)
-
-    net.connect("Generate.OUT", "Subnet.IN")
-    net.connect("Subnet.OUT", "Discard.IN")
-
-    net.go()
-
-    assert dis.values == [
-        '', '000005', '000004', '000003', '000002', '000001', '',
-    ]
-
-
 def test_export_serialization():
     net = Network()
 
@@ -1016,6 +956,7 @@ def test_network_apply():
     })
 
     assert outputs['OUT'] == 10
+
 
 def test_network_apply_with_outputs():
     net = Network()
